Create separate profiles for lambda projects
