+++ src/main/kotlin/com/jetbrains/rider/aspire/sessionHost/awsLambda/LambdaProjectSessionRunProfile.kt
@@ -1,16 +1,16 @@
-package com.jetbrains.rider.aspire.sessionHost.dotnetProject
+package com.jetbrains.rider.aspire.sessionHost.awsLambda
 
 import com.intellij.execution.Executor
 import com.intellij.execution.process.ProcessListener
 import com.intellij.execution.runners.ExecutionEnvironment
 import com.jetbrains.rd.util.lifetime.Lifetime
-import com.jetbrains.rider.aspire.sessionHost.projectLaunchers.DotNetExecutableProjectSessionDebugProfileState
+import com.jetbrains.rider.aspire.sessionHost.projectLaunchers.DotNetExecutableSessionRunProfileState
 import com.jetbrains.rider.aspire.sessionHost.projectLaunchers.ProjectSessionProfile
 import com.jetbrains.rider.runtime.DotNetExecutable
 import com.jetbrains.rider.runtime.dotNetCore.DotNetCoreRuntime
 import java.nio.file.Path
 
-class DotNetProjectSessionDebugProfile(
+class LambdaProjectSessionRunProfile(
     private val sessionId: String,
     projectPath: Path,
     dotnetExecutable: DotNetExecutable,
@@ -18,11 +18,11 @@ class DotNetProjectSessionDebugProfile(
     private val sessionProcessEventListener: ProcessListener,
     private val sessionProcessLifetime: Lifetime,
     aspireHostProjectPath: Path?
-) : ProjectSessionProfile(projectPath, dotnetExecutable, aspireHostProjectPath, true) {
+) : ProjectSessionProfile(projectPath, dotnetExecutable, aspireHostProjectPath, false) {
     override fun getState(
         executor: Executor,
         environment: ExecutionEnvironment
-    ) = DotNetExecutableProjectSessionDebugProfileState(
+    ) = DotNetExecutableSessionRunProfileState(
         sessionId,
         dotnetExecutable,
         dotnetRuntime,
